requirements/local.txt
-r base.txt  # includes the base.txt requirements file

coverage==4.3.4
django-debug-toolbar==1.6
selenium==3.0.2
